+++ startos/versions/index.ts
@@ -3,8 +3,9 @@ import { current } from './current'
 import { v0_2_0 } from './v0_2_0'
 import { v0_3_0 } from './v0_3_0'
 import { v0_3_1 } from './v0_3_1'
+import { v0_4_0 } from './v0_4_0'
 
 export const versionGraph = VersionGraph.of({
   current,
-  other: [v0_2_0, v0_3_0, v0_3_1],
+  other: [v0_2_0, v0_3_0, v0_3_1, v0_4_0],
 })